working on search bar functionality
