+++ Client/src/app/Core/ToolBars/Outside/outside-toolbar.component.html
@@ -4,4 +4,4 @@
         <login></login>
     </mat-toolbar-row>
 </mat-toolbar>
-<mat-progress-bar mode="buffer" *ngIf="getIsLoading()"></mat-progress-bar> 
+<mat-progress-bar class="progress-bar" mode="buffer" *ngIf="getIsLoading()"></mat-progress-bar>
