+++ Client/src/app/Core/ToolBars/Outside/outside-toolbar.component.scss
@@ -1,3 +1,5 @@
+@import '../../../../theme.scss';
+
 .toolbar-container {
     width: 100%;
     mat-toolbar-row {
@@ -14,6 +16,10 @@
             font-size: 30px;
         }
     }
+}
+
+.progress-bar {
+    background-color: $accent;
 }
 
 @media screen and (max-width: 750px) {
