+++ Client/src/app/Main/Components/Dashboard/dashboard.component.html
@@ -1,19 +1,19 @@
 <div id="dashboard" *ngIf="!isError(error)">
     <mat-card class="tab-container">
         <mat-tab-group mat-align-tabs="center" animationDuration="0ms" color="primary">
-            <mat-tab label="Recent">
-                <infinity-scroll
-                    [isGuest]="isGuest"
-                    [movies]="searchedMovies"
-                    [typeScrolled]="'searchHistory'">
-                </infinity-scroll>
-            </mat-tab>
             <mat-tab label="Favorites">
                 <infinity-scroll
                     [isGuest]="isGuest"
                     [movies]="favorites"
                     [typeScrolled]="'favorites'"
                     (getNextMovies)="onGetNextMovies($event)">
+                </infinity-scroll>
+            </mat-tab>
+            <mat-tab label="History">
+                <infinity-scroll
+                    [isGuest]="isGuest"
+                    [movies]="searchedMovies"
+                    [typeScrolled]="'searchHistory'">
                 </infinity-scroll>
             </mat-tab>
         </mat-tab-group>
