+++ Client/src/app/Main/Components/Search/search.component.html
@@ -1,0 +1,34 @@
+<div id="search">
+    <div class="search-options">
+        <mat-form-field class="input-box">
+            <mat-label>Search Movies...</mat-label>
+            <input  
+                #imdbIdSearch
+                type="text"
+                matInput 
+                [(ngModel)]="search"  
+                (keydown)="clearSearchResults()"
+                [disabled]="gettingMovie"
+                autocomplete="off">
+            <mat-icon class="search-icon" (click)="searchMovies(search, year)">search</mat-icon>
+        </mat-form-field>
+        <mat-form-field class="year-box">
+            <mat-label>Year</mat-label>
+            <mat-select [(value)]="year" class="year-input">
+                <mat-option>None</mat-option>
+                <mat-option value="{{yr}}" *ngFor="let yr of getYears()" (click)="searchMovies(search, year)">
+                    {{yr}}
+                </mat-option>
+            </mat-select>
+        </mat-form-field>
+    </div>
+</div>
+<div class="search-container" *ngIf="moviesExist(imdbs, search)">
+    <div class="search-results" *ngFor="let imdb of imdbs" (click)="getOrCreateMovie(imdb)">
+        <span class="title">{{imdb.title}}</span> 
+        <span class="year">{{imdb.year}}</span>
+    </div>
+</div>
+<div *ngIf="noSearchResults" class="empty-search">
+    No movies found from search
+</div>
