+++ Client/src/app/Main/Components/Search/search.component.scss
@@ -1,0 +1,199 @@
+@import '../../../../theme.scss';
+
+#search {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 5px;
+
+    ::ng-deep .mat-tab-body-content {
+        display: flex;
+    }
+
+    .search-options {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        flex-direction: row;
+        z-index: 1000;
+        .input-box {
+            line-height: unset;
+            margin-right: 5px;
+            ::ng-deep .mat-form-field-wrapper  {
+                height: 20px;
+                width: 550px;
+                background-color: $ternary;
+                border: 2px solid $accent;
+                border-radius: 25px;
+
+                ::ng-deep .mat-form-field-infix {
+                    display: flex;
+                    align-items: center;
+                    justify-content: center;
+                    flex-direction: row;
+                    border: none;
+                    padding: unset;
+
+                    ::ng-deep .mat-input-element {
+                        font-size: 25px;
+                        caret-color: $primary;
+                        color: $primary;
+                        padding-left: 10px;
+                    }
+                    ::ng-deep .mat-form-field-label-wrapper {
+                        top: -1.04375em;
+                        ::ng-deep .mat-form-field-label {
+                            top: 1.18125em;
+                            font-size: 25px;
+                            font-weight: 700;
+                            color: $primary;
+                            padding-left: 5px;
+                        }
+                    }
+                }
+
+                ::ng-deep .mat-form-field-underline {
+                    display: none;
+                }
+            }
+
+            .search-icon {
+                height: auto;
+                width: auto;
+                font-size: 50px;
+                color: $primary;
+            }
+    
+            .search-icon:hover {
+                cursor: pointer;
+            }
+        }    
+    
+        .year-box {           
+            line-height: unset;
+            ::ng-deep .mat-form-field-wrapper {
+                height: 20px;
+                width: 100px;
+                background-color: $ternary;
+                border: 2px solid $accent;
+                border-radius: 25px;
+
+                ::ng-deep .mat-form-field-infix {
+                    display: flex;
+                    align-items: center;
+                    justify-content: center;
+                    flex-direction: row;
+                    border: none;
+                    padding: unset;
+                    
+                    .year-input {
+                        box-shadow: none;
+                        font-size: 15px;
+                        background-color: $ternary;
+                        border-radius: 25px;
+        
+                        ::ng-deep .mat-select-trigger {
+                            height: 3.125em;
+
+                            ::ng-deep .mat-select-value {
+                                font-size: 25px;
+                                           
+                                ::ng-deep .mat-select-value-text {
+                                    color: $ternary!important;
+                                }
+                            }
+            
+                            ::ng-deep .mat-select-arrow-wrapper {
+                                padding-right: 5px;
+                                ::ng-deep .mat-select-arrow {
+                                    color: $primary;
+                                }
+                            }
+                        }
+                    }
+                    ::ng-deep .mat-form-field-label-wrapper {
+                        top: -1.04375em;
+                        ::ng-deep .mat-form-field-label {
+                            font-size: 25px;
+                            padding-left: 5px;
+                            color: $primary !important;
+                            font-weight: 700;
+                            top: 1.18125em;
+                        }
+                    }
+                }
+
+                ::ng-deep .mat-form-field-underline {
+                    display: none;
+                }
+            }
+        } 
+    }
+}
+.search-container {
+    position: absolute;
+    top: 64px;
+    width: 663px;
+    .search-results {
+        background-color: $ternary;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        border-bottom: 1px black solid;
+        .title, .year {
+            color: $primary;
+            padding: 5px;
+        }
+        .title {
+            width: 75%;
+        }
+    }
+    .search-results:hover {
+        cursor: pointer;
+        background-color: $accent;
+    }
+    table {
+        width: 663px;
+        background-color: #b0bec5;
+        .mat-header-row {
+            .mat-header-cell {
+                font-size: 20px;
+                font-weight: 900;
+                color: white;
+                background-color: #c2185b;
+            }
+        }
+        .mat-row {
+            height: 60px;
+            .mat-cell {
+                color: #040c24;
+                font-size: 20px;
+                font-weight: 600;
+            }
+            .mat-column-Title {
+                width: 375px;
+            }
+            .mat-column-Year {
+                width: 50px;
+                padding: 5px;
+            }
+        }
+        .mat-row:hover {
+            background-color: rgba(129, 122, 123, 0.507);
+            cursor: pointer;
+        }
+    }
+    .mat-paginator {
+        background-color: #c2185b;
+        ::ng-deep .mat-paginator-range-actions {
+            font-size: 20px;
+            color: white;
+        }
+    }
+}
+.empty-search {
+    font-size: 25px;
+    font-weight: bold;
+    padding: 20px;
+    color: #040c24;
+}
